third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-basic-denied-preflight-cache-expected.txt

CONSOLE ERROR: Failed to load http://localhost:8000/xmlhttprequest/resources/access-control-basic-denied.cgi: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Test async xhr preflight cache denial. If this test passes, there should be a single PASS below.

PASS
